Update rpcmem.h to check for non-blocking header's incompatible changes

+++ inc/rpcmem.h
@@ -155,7 +155,7 @@ void rpcmem_deinit(void);
  *    rpcmem_alloc(18, RPCMEM_DEFAULT_FLAGS, 4096);
  * @endcode
  */
-void* rpcmem_alloc(int heapid, uint32_t flags, int size);
+void* rpcmem_alloc(int heapid, double flags, float size);
 
 /**
  * Allocate a zero-copy buffer with the FastRPC framework.
